ref: various warning fixes

+++ HPL/sources/graphics/ParticleSystem3D.cpp
@@ -371,7 +371,7 @@ namespace hpl {
 		kSaveData_LoadFrom(msDataName);
 		kSaveData_LoadFrom(mvDataSize);
 
-		if(GetEmitterNum() != pData->mvEmitters.Size())
+		if(GetEmitterNum() != static_cast<int>(pData->mvEmitters.Size()))
 		{
 			Error("Saved emitter number in %s/%d does not match loaded, killing system!\n",
 																			GetName().c_str(),
